+++ Assignments/A2MinhNB_HE191060_SE1941JV_NMS_BE/src/main/java/A2MinhNB_HE191060/SE1941JV/dto/account/AccountResponse.java
@@ -10,9 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AccountResponse {
-    private Long accountId;
+    private Integer accountId;
     private String accountName;
     private String accountEmail;
     private Integer accountRole;
-    // never expose password
 }
