+++ Assignments/A2MinhNB_HE191060_SE1941JV_NMS_BE/src/main/java/A2MinhNB_HE191060/SE1941JV/security/AccountUserDetails.java
@@ -49,7 +49,7 @@ public class AccountUserDetails implements UserDetails {
         return true;
     }
 
-    public Long getAccountId() {
+    public Integer getAccountId() {
         return account.getAccountId();
     }
